backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
openai
xai-sdk
langchain
langchain-community
chromadb
sentence-transformers
psycopg2-binary
pgvector
numpy
tiktoken
python-multipart
PyJWT
pytest
httpx
